Small changes in builder and director
