+++ src/Service/AmoApiAuthDirector.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 namespace App\Service;
 
 use AmoCRM\Client\AmoCRMApiClient;
-use Symfony\Component\Dotenv\Dotenv;
 
 class AmoApiAuthDirector
 {
@@ -17,17 +16,6 @@ class AmoApiAuthDirector
         $builder = new AmoApiAuthBuilder($apiClientConfig);
         $this->builder = $builder;
         $this->tokenPath = $tokenPath;
-    }
-
-    public static function getDefaultCredentials(): array
-    {
-        (new Dotenv())->usePutenv()->bootEnv(dirname(__DIR__).'/../.env');
-        return [
-            'redirect_uri' => getenv('REDIRECT_URI'),
-            'client_id' => getenv('INTEGRATION_ID'),
-            'client_secret' => getenv('SECRET'),
-            'base_domain' => getenv('BASE_DOMAIN')
-        ];
     }
 
     public function buildAuthentication(): self
